Made a small change to GUIManager
Your changes made it so that if someone hits the return key once it will
only run what's in the if once but if you keep holding down return it
will start calling the method repeatedly so I added a couple of lines to
prevent this. If you have any questions about the lines of code I added
let me know.